Refactor CallForwardingHandler
The changes here should be separated out in to a series of commits.
However, as it was a lot of experimentation, and I forgot to be as
diligent as I normally would, that's not practical right now.

So, here's what's changed. The autoload namespace has been updated to
simplify it. As it was previously, it was too long-winded; unnecessarily
complicated. Secondly, the code now injects the current date to use as a
base point of reference into the CallForwardingHandler making it easier
to test. This is handled via PSR-20 through the psr/clock package which
provides ClockInterface. Thirdly, it fixes a bug in the comparison
logic, where the current day and time is compared to the work week.